feat/customer_feature сделан круд для клиента, добавлена валидация номера телефона

+++ services/base_service.py
@@ -3,7 +3,6 @@ from dataclasses import dataclass
 from typing import Generic, Any
 from uuid import UUID
 
-
 from core.db.generics import ModelType, SchemaType
 from repository.base_repository import RepositoryBase, RepositoryORM
 
@@ -47,7 +46,8 @@ class MainServices(BaseServices, Generic[ModelType, SchemaType]):
 
     async def delete(self, id_: UUID) -> MODEL | None:
         obj = await self.repository.get(id_=id_)
-        await self.repository.delete(model=obj)
+        if obj:
+            await self.repository.delete(model = obj)
         return obj
 
     async def patch(self, id_: UUID, data: dict[str, Any]) -> MODEL:
